add optimisations for compile speed
