+++ just_enough_mod/Cargo.toml
@@ -8,5 +8,5 @@ license = "MIT"
 
 [dependencies]
 just_enough_mod_core = { path = "../just_enough_mod_core" }
-bevy = { version = "0.12", features = [ "wayland" ] }
+bevy = { version = "0.12", features = [ "wayland", "dynamic_linking" ] }
 libloading = "0.8.1"
