+++ just_enough_mod_base/src/lib.rs
@@ -8,11 +8,24 @@ impl Plugin for JustEnoughModBase {
         println!("JustEnoughModBase::init")
     }
 
-    fn bevy_init(&self, sched: Option<&mut Schedule>) {
-        sched.expect("shoudl be here").add_systems(rotate);
-        // app
-        // .add_systems(Update, rotate);
+    fn bevy_init(&self, app: &mut App) {
+        // app.get_schedule_mut(Update).expect("shoudl be here").add_systems(rotate);
+        app
+        .add_systems(Update, rotate)
+        .add_systems(Startup, setup);
+
     }
+}
+
+fn setup(mut commands: Commands, asset_server: Res<AssetServer>) {
+    // Camera
+    commands.spawn(Camera2dBundle::default());
+
+    // Sprite
+    commands.spawn(SpriteBundle {
+        texture: asset_server.load("bevy.png"),
+        ..default()
+    });
 }
 
 fn rotate(mut query: Query<&mut Transform, With<Sprite>>, time: Res<Time>) {
